move event state machine to event package
